Share one HAPI endpoint construct between both stacks
The forms-server and ehr-proxy stacks each carried their own copy of the
HapiEndpoint construct. The two were already near-identical, differing only in
container name, task size and one extra HAPI setting, and adding the RDS
database to both widened the duplication to include the engine version, image
tag, storage, backup and credential wiring. Keeping the persistence
configuration correct in two places is exactly the kind of drift that produced
the earlier split between HAPI 8.4.0 and 8.10.0.

Replace both copies with a single shared-hapi-endpoint package, parameterised on
the few values that genuinely differ. The image tag and PostgreSQL version are
exported as constants so the two deployments cannot drift apart again, and the
caller-supplied HAPI settings are merged underneath the datasource settings so a
caller can extend the configuration but cannot unset the datasource.

Construct ids are passed in through constructIdPrefix rather than derived from
the construct's own id. CloudFormation logical ids come from the construct path,
and a renamed resource is a replaced resource, which for the databases would
mean losing their contents. Passing the prefix explicitly keeps every logical id
byte-identical to what is deployed today.

This is a pure refactor. Both stacks were synthesized before and after, and the
resulting CloudFormation templates are byte-identical, so cdk diff reports no
change.
