+++ deployment/ehr-proxy/ehr-proxy-app/package.json
@@ -22,8 +22,8 @@
   "dependencies": {
     "aws-cdk-lib": "^2.260.0",
     "constructs": "^10.6.0",
-    "ehr-proxy-hapi-endpoint": "^0.1.0",
     "ehr-proxy-smart-proxy": "^0.1.0",
+    "shared-hapi-endpoint": "^0.1.0",
     "source-map-support": "^0.5.21"
   }
 }
