--- deployment/ehr-proxy/hapi-endpoint/README.md
@@ -1,12 +1,0 @@
-# Welcome to your CDK TypeScript Construct Library project
-
-You should explore the contents of this project. It demonstrates a CDK Construct Library that includes a construct (`HapiEndpoint`)
-which contains an Amazon SQS queue that is subscribed to an Amazon SNS topic.
-
-The construct defines an interface (`HapiEndpointProps`) to configure the visibility timeout of the queue.
-
-## Useful commands
-
-* `npm run build`   compile typescript to js
-* `npm run watch`   watch for changes and compile
-* `npm run test`    perform the jest unit tests
