--- deployment/forms-server/hapi-endpoint/.gitignore
@@ -1,8 +1,0 @@
-*.js
-!jest.config.js
-*.d.ts
-node_modules
-
-# CDK asset staging directory
-.cdk.staging
-cdk.out
