--- deployment/forms-server/hapi-endpoint/.npmignore
@@ -1,6 +1,0 @@
-*.ts
-!*.d.ts
-
-# CDK asset staging directory
-.cdk.staging
-cdk.out
