--- deployment/forms-server/hapi-endpoint/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "forms-server-hapi-endpoint",
-  "version": "0.1.0",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "scripts": {
-    "build": "tsc",
-    "watch": "tsc -w",
-    "test": "jest"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.12",
-    "@types/node": "22.15.30",
-    "aws-cdk-lib": "^2.260.0",
-    "constructs": "^10.6.0",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.4.2",
-    "typescript": "~5.9.3"
-  },
-  "peerDependencies": {
-    "aws-cdk-lib": "^2.260.0",
-    "constructs": "^10.3.0"
-  },
-  "dependencies": {
-    "aws-cdk": "^2.1031.0"
-  }
-}
