+++ deployment/shared/hapi-endpoint/.npmignore
